requirements.txt
azure-core==1.10.0
azure-functions==1.6.0
azure-storage-blob==12.7.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
Flask==1.1.2
Flask-Cors==3.0.10
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msrest==0.6.19
oauthlib==3.1.0
pycparser==2.20
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
